working printing ordering with sortpointers.c

+++ outputarray.c
@@ -18,9 +18,9 @@
 // }
 #include <stdio.h>
 extern "C"{
-  void output_array(int* arr[], int size){
+  void output_array(double* arr[], int size){
     for (int i = 0; i < size;i++ ){
-      printf("%d \n",(*arr[i]));
+      printf("%lf \n",(*arr[i]));
     }
   }
 }
